requirements.txt
coloredlogs
matplotlib
numpy
Pillow
torch==1.13.1+cu117 
torchvision==0.14.1+cu117 
torchaudio==0.13.1
tqdm
opencv-python